modify(spring-boot): Increase code quality
