+++ src/main/java/com/apcsa/practice/config/SecurityConfig.java
@@ -1,4 +1,3 @@
-// src/main/java/com/apcsa/practice/config/SecurityConfig.java
 package com.apcsa.practice.config;
 
 import org.springframework.context.annotation.Bean;
@@ -7,19 +6,30 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.web.SecurityFilterChain;
 
+/**
+ * Spring Security Configuration Class
+ * This class configures the security settings for the application,
+ * defining access rules for different endpoints and security features.
+ */
 @Configuration
 @EnableWebSecurity
 public class SecurityConfig {
 
+    /**
+     * Configures the security filter chain for HTTP requests.
+     * 
+     * @param http The HttpSecurity object to configure
+     * @return The built SecurityFilterChain
+     * @throws Exception If an error occurs during configuration
+     */
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         http
             .authorizeHttpRequests(authz -> authz
-                .requestMatchers("/api/**").permitAll()  // 允許所有 /api/** 路徑
-                .anyRequest().authenticated()
+                .requestMatchers("/api/**").permitAll()  // Allow unrestricted access to all /api/** paths
+                .anyRequest().authenticated()            // Require authentication for any other request
             )
-            .csrf(csrf -> csrf.disable());  // 暫時關閉 CSRF 保護
-        
+            .csrf(csrf -> csrf.disable());  // Temporarily disable CSRF protection for development purposes
         return http.build();
     }
 }
