+++ src/main/java/com/apcsa/practice/model/Question.java
@@ -1,0 +1,246 @@
+package com.apcsa.practice.model;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+/**
+ * Question entity class representing an AP CSA practice question.
+ * This class maps to documents in the "questions" collection in MongoDB.
+ */
+@Document(collection = "questions")
+public class Question {
+    
+    /**
+     * Unique identifier for the question, automatically generated by MongoDB.
+     */
+    @Id
+    private String id;
+    
+    /**
+     * Sequential number assigned to the question for ordering purposes.
+     */
+    private int questionNumber;
+    
+    /**
+     * The text content of the question.
+     */
+    private String questionText;
+    
+    /**
+     * First answer option (A).
+     */
+    private String optionA;
+    
+    /**
+     * Second answer option (B).
+     */
+    private String optionB;
+    
+    /**
+     * Third answer option (C).
+     */
+    private String optionC;
+    
+    /**
+     * Fourth answer option (D).
+     */
+    private String optionD;
+    
+    /**
+     * The correct answer, represented as "A", "B", "C", or "D".
+     */
+    private String correctAnswer;
+    
+    /**
+     * Default no-argument constructor required by MongoDB.
+     */
+    public Question() {
+    }
+    
+    /**
+     * Parameterized constructor to create a new Question with all required fields.
+     * 
+     * @param questionNumber The sequential number of the question
+     * @param questionText The text content of the question
+     * @param optionA First answer option (A)
+     * @param optionB Second answer option (B)
+     * @param optionC Third answer option (C)
+     * @param optionD Fourth answer option (D)
+     * @param correctAnswer The correct answer ("A", "B", "C", or "D")
+     */
+    public Question(int questionNumber, String questionText, String optionA, String optionB, 
+                        String optionC, String optionD, String correctAnswer) {
+        this.questionNumber = questionNumber;
+        this.questionText = questionText;
+        this.optionA = optionA;
+        this.optionB = optionB;
+        this.optionC = optionC;
+        this.optionD = optionD;
+        this.correctAnswer = correctAnswer;
+    }
+    
+    // Getters and Setters
+    
+    /**
+     * Gets the unique identifier of the question.
+     * 
+     * @return The question's ID
+     */
+    public String getId() {
+        return id;
+    }
+    
+    /**
+     * Sets the unique identifier of the question.
+     * 
+     * @param id The question's ID to set
+     */
+    public void setId(String id) {
+        this.id = id;
+    }
+    
+    /**
+     * Gets the sequential number of the question.
+     * 
+     * @return The question number
+     */
+    public int getQuestionNumber() {
+        return questionNumber;
+    }
+    
+    /**
+     * Sets the sequential number of the question.
+     * 
+     * @param questionNumber The question number to set
+     */
+    public void setQuestionNumber(int questionNumber) {
+        this.questionNumber = questionNumber;
+    }
+    
+    /**
+     * Gets the text content of the question.
+     * 
+     * @return The question text
+     */
+    public String getQuestionText() {
+        return questionText;
+    }
+    
+    /**
+     * Sets the text content of the question.
+     * 
+     * @param questionText The question text to set
+     */
+    public void setQuestionText(String questionText) {
+        this.questionText = questionText;
+    }
+    
+    /**
+     * Gets the first answer option (A).
+     * 
+     * @return Option A text
+     */
+    public String getOptionA() {
+        return optionA;
+    }
+    
+    /**
+     * Sets the first answer option (A).
+     * 
+     * @param optionA Option A text to set
+     */
+    public void setOptionA(String optionA) {
+        this.optionA = optionA;
+    }
+    
+    /**
+     * Gets the second answer option (B).
+     * 
+     * @return Option B text
+     */
+    public String getOptionB() {
+        return optionB;
+    }
+    
+    /**
+     * Sets the second answer option (B).
+     * 
+     * @param optionB Option B text to set
+     */
+    public void setOptionB(String optionB) {
+        this.optionB = optionB;
+    }
+    
+    /**
+     * Gets the third answer option (C).
+     * 
+     * @return Option C text
+     */
+    public String getOptionC() {
+        return optionC;
+    }
+    
+    /**
+     * Sets the third answer option (C).
+     * 
+     * @param optionC Option C text to set
+     */
+    public void setOptionC(String optionC) {
+        this.optionC = optionC;
+    }
+    
+    /**
+     * Gets the fourth answer option (D).
+     * 
+     * @return Option D text
+     */
+    public String getOptionD() {
+        return optionD;
+    }
+    
+    /**
+     * Sets the fourth answer option (D).
+     * 
+     * @param optionD Option D text to set
+     */
+    public void setOptionD(String optionD) {
+        this.optionD = optionD;
+    }
+    
+    /**
+     * Gets the correct answer for the question.
+     * 
+     * @return The correct answer ("A", "B", "C", or "D")
+     */
+    public String getCorrectAnswer() {
+        return correctAnswer;
+    }
+    
+    /**
+     * Sets the correct answer for the question.
+     * 
+     * @param correctAnswer The correct answer to set ("A", "B", "C", or "D")
+     */
+    public void setCorrectAnswer(String correctAnswer) {
+        this.correctAnswer = correctAnswer;
+    }
+    
+    /**
+     * Returns a string representation of the Question object.
+     * 
+     * @return A string containing all question properties
+     */
+    @Override
+    public String toString() {
+        return "Question{" +
+                "id='" + id + '\'' +
+                ", questionNumber=" + questionNumber +
+                ", questionText='" + questionText + '\'' +
+                ", optionA='" + optionA + '\'' +
+                ", optionB='" + optionB + '\'' +
+                ", optionC='" + optionC + '\'' +
+                ", optionD='" + optionD + '\'' +
+                ", correctAnswer='" + correctAnswer + '\'' +
+                '}';
+    }
+}
