+++ src/main/resources/application.properties
@@ -1,7 +1,17 @@
+# Application name - identifies this application in Spring ecosystem
 spring.application.name=ap-csa-practice
 
 # MongoDB Configuration
-# Connection URI will be used instead of host/port
-# If your MongoDB requires authentication, include username and password in the URI, for example:
-# mongodb://username:password@host:port/database
+# =====================
+# Connection URI is used instead of individual host/port settings
+# Format: mongodb://[username:password@]host[:port]/database
+# The URI below uses an environment variable with a fallback to localhost if not set
+# For development: mongodb://localhost:27017/apcsadb
+# For production: Set the SPRING_DATA_MONGODB_URI environment variable with appropriate credentials
 spring.data.mongodb.uri=${SPRING_DATA_MONGODB_URI:mongodb://localhost:27017/apcsadb}
+
+# Note: For production deployments, it's recommended to:
+# 1. Use environment variables for sensitive configuration
+# 2. Enable authentication for MongoDB
+# 3. Configure network security to restrict database access
+# 4. Consider using MongoDB Atlas or other managed services for production databases
