+++ src/test/resources/application.properties
@@ -1,8 +1,26 @@
+# Test-specific application properties
+# These settings are only used during test execution and override the main application.properties
+
+# Application name - identifies this application in Spring ecosystem
 spring.application.name=ap-csa-practice
 
-# MongoDB Configuration for tests
-# This will be overridden by environment variables in Docker
+# MongoDB Configuration for Test Environment
+# ==========================================
+# Uses a separate database for tests to avoid affecting development or production data
+# The URI uses an environment variable with a fallback to a local test database
+# Note: The '_test' suffix in the database name helps distinguish it from the development database
 spring.data.mongodb.uri=${SPRING_DATA_MONGODB_URI:mongodb://localhost:27017/apcsadb_test}
 
-# Disable auto-index creation during tests to improve startup time
+# Performance Optimization for Tests
+# =================================
+# Disable automatic index creation during tests to improve startup time
+# This is safe for tests since we don't rely on custom indexes for test execution
 spring.data.mongodb.auto-index-creation=false
+
+# Additional Test-Specific Settings
+# ================================
+# You can add more test-specific configurations here, such as:
+# - Reduced logging levels
+# - In-memory database configurations
+# - Mock service endpoints
+# - Disabled security features
